refactor(parser): split class member method construction
